changed to newer gymnasium interface: This splits the "done" signal into two subsignals - terminated - truncated the idea is that for bootstrapping you should only consider the "terminated" as a stop for the bootstrap, e.g. V(s) = R(s) + gamma*(1-terminated)*V(s') i.e. "terminated" actually ends the episode, while "truncated" only tells you the method stopped due to e.g. time limits. In the trunacted style you would typically continue your bootstrapping!
Since we only have a time limit no failure condition (yet), this is currently
implemented as only using truncation.
If you want to recover the old behaviour, just look at done = terminated or truncated
or use the "gymnasium to gym" wrapper.

A small additional change is that everything is now a relative import,
which helps if you have the env in its own folder